11.03.2023

+++ BACKEND/src/main/resources/static/js/view-post-script.js
@@ -193,6 +193,6 @@ function printComments(data) {
 function correctDate(date) {
     let data = new Date(date.toString());
     return data.toLocaleDateString('en-GB', {
-        day: 'numeric', month: 'long', year: 'numeric'
+        day: 'numeric', month: 'short', year: 'numeric'
     }).replace(/ /g, ' ');
 }
